Big update of mainly graphics and settings management
Easier to scale graphics depending on the initial window size.
Most of the scenes are adapted to this now.

New fish sprites. Waves! Bubbles! Gradients here and there.

Settings for a lot of things are stored in one place and can be
altered through a GUI. For now, the settings are saved to a cookie,
if possible (not possible in Chrome on localhost).